move helper functions into validators
